# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/app/v1/shape": {
            "get": {
                "description": "Client uses this API to get all created shapes.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ShapeApi"
                ],
                "summary": "Get All Shapes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.AllShapeInfoResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/app/v1/shape/calculate": {
            "post": {
                "description": "Client uses this API to Calculate Area or Perimeter of a specific Shape based on ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ShapeApi"
                ],
                "summary": "Calculate Area or Perimeter of a specific Shape based on ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "GaphQL Query",
                        "name": "ShapeCalculationBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.ShapeCalculationBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.ShapeCalculationResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/app/v1/shape/create": {
            "post": {
                "description": "Client uses this API to create a shape.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ShapeApi"
                ],
                "summary": "Create Shape",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Shape Information",
                        "name": "ShapeInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.ShapeInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.ShapeCreateResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/app/v1/shape/{shape_id}": {
            "get": {
                "description": "Client uses this API to get a created shape with specific shape ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ShapeApi"
                ],
                "summary": "Get Specific Shape From ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.ShapeInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Client uses this API to update shape's edge values.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ShapeApi"
                ],
                "summary": "Update a specific shape",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Shape's Edge Information",
                        "name": "EdgeInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.EdgeInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.OKResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Client uses this API to delete a shape specified by ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ShapeApi"
                ],
                "summary": "Delete a specific shape",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.OKResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/app/v1/user/login": {
            "post": {
                "description": "User uses this API to login the system.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserApi"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "User Login Body",
                        "name": "LoginInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.LoginInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.LoginSuccessResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/v1.UnauthorizedResponse"
                        }
                    }
                }
            }
        },
        "/app/v1/user/signup": {
            "post": {
                "description": "Client uses this API to create a new user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserApi"
                ],
                "summary": "Signup",
                "parameters": [
                    {
                        "description": "User Signup Body",
                        "name": "SignUpInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.SignupInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.OKResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "v1.AllShapeInfoResp": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.ShapeItem"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                }
            }
        },
        "v1.EdgeInfo": {
            "type": "object",
            "properties": {
                "edges": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "['1'",
                        " '2'",
                        " '3']"
                    ]
                }
            }
        },
        "v1.ErrorResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "An error description"
                }
            }
        },
        "v1.LoginInfo": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "passw0rd"
                },
                "username": {
                    "type": "string",
                    "example": "peterweb"
                }
            }
        },
        "v1.LoginSuccessResp": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/v1.TokenResp"
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                }
            }
        },
        "v1.OKResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "OK"
                }
            }
        },
        "v1.ShapeCalculationBody": {
            "type": "object",
            "properties": {
                "query": {
                    "type": "string",
                    "example": "{area(shape_id:1)\n perimeter(shape_id:2)}"
                }
            }
        },
        "v1.ShapeCalculationResp": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/v1.ShapeValue"
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                }
            }
        },
        "v1.ShapeCreateResp": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/v1.ShapeId"
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                }
            }
        },
        "v1.ShapeId": {
            "type": "object",
            "properties": {
                "shape_id": {
                    "type": "integer",
                    "example": 123
                }
            }
        },
        "v1.ShapeInfo": {
            "type": "object",
            "properties": {
                "edges": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "['1'",
                        " '2'",
                        " '3']"
                    ]
                },
                "shape": {
                    "type": "string",
                    "example": "triangle"
                }
            }
        },
        "v1.ShapeItem": {
            "type": "object",
            "properties": {
                "edges": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "['1'",
                        " '2'",
                        " '3']"
                    ]
                },
                "shape": {
                    "type": "string",
                    "example": "triangle"
                },
                "shape_id": {
                    "type": "integer",
                    "example": 123
                }
            }
        },
        "v1.ShapeValue": {
            "type": "object",
            "properties": {
                "area": {
                    "type": "string",
                    "example": "12.1234"
                },
                "perimeter": {
                    "type": "string",
                    "example": "30.0000"
                }
            }
        },
        "v1.SignupInfo": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "passw0rd"
                },
                "username": {
                    "type": "string",
                    "example": "peterweb"
                }
            }
        },
        "v1.TokenResp": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "abx...xyz"
                }
            }
        },
        "v1.UnauthorizedResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "Unauthorized"
                }
            }
        }
    }
}